Struct maa_framework::instance::MaaInstance

pub struct MaaInstance<T> { /* private fields */ }

The MaaInstance struct is the main entry point for the Maa library.

It is used to create and manage the Maa instance for running tasks.

Example

use maa_framework::instance::MaaInstance;

let instance = MaaInstance::new(None);
// let param = serde_json::json!({"param": "value"});
// instance.post_task("task", param);

Note

MaaInstance,MaaResourceInstance and MaaControllerInstance use the same mechanism to manage the lifetime of the underlying C++ object. That is, if the object is created from the Rust code (like MaaInstance::new), the object will be destroyed when it goes out of scope. In this case, it is your responsibility to ensure that the object is not used after it has been destroyed. If the object is created from the C++ code, then you will not have to worry about the lifetime of the object.

Implementations

impl<T> MaaInstance<T>

pub fn new(handler: Option<T>) -> Self
where T: CallbackHandler,

Create a new MaaInstance.

Parameters

handler - An optional callback handler for handling Maa events.

pub fn set_option(&self, option: MaaInstOption) -> MaaResult<()>

pub fn bind_resource(&self, res: &MaaResourceInstance<T>) -> MaaResult<()>

pub fn bind_controller( &self, controller: &MaaControllerInstance<T> ) -> MaaResult<()>

pub fn inited(&self) -> bool

pub fn post_task<P>(&self, entry: &str, param: P) -> MaaTaskId
where P: TaskParam,

pub fn set_task_param(&self, task_id: MaaTaskId, param: &str) -> MaaResult<()>

pub fn task_status(&self, task_id: MaaTaskId) -> MaaResult<MaaStatus>

pub fn wait_task(&self, task_id: MaaTaskId) -> MaaResult<MaaStatus>

pub fn task_all_finished(&self) -> bool

pub fn post_stop(&self) -> MaaResult<()>

pub fn stop(&self) -> MaaResult<()>

pub fn resource(&self) -> MaaResourceInstance<T>

pub fn controller(&self) -> MaaControllerInstance<T>

pub fn register_custom_recognizer<R>( &mut self, name: &str, recognizer: R ) -> MaaResult<()>
where R: MaaCustomRecognizer,

pub fn unregister_custom_recognizer(&mut self, name: &str) -> MaaResult<()>

pub fn clear_custom_recognizers(&mut self) -> MaaResult<()>

pub fn register_custom_action<A>( &mut self, name: &str, action: A ) -> MaaResult<()>
where A: MaaCustomAction,

pub fn unregister_custom_action(&mut self, name: &str) -> MaaResult<()>

pub fn clear_custom_actions(&mut self) -> MaaResult<()>